Added support for h2 DB.

+++ src/main/java/sawfowl/regionguard/configure/storage/MySqlStorage.java
@@ -1,16 +1,13 @@
 package sawfowl.regionguard.configure.storage;
 
-import java.io.BufferedReader;
-import java.io.BufferedWriter;
-import java.io.StringReader;
-import java.io.StringWriter;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
+import java.util.Collection;
 import java.util.HashMap;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
 import java.util.concurrent.TimeUnit;
@@ -23,44 +20,28 @@ import org.spongepowered.api.event.lifecycle.RefreshGameEvent;
 import org.spongepowered.api.scheduler.ScheduledTask;
 import org.spongepowered.api.scheduler.Task;
 import org.spongepowered.api.world.server.ServerWorld;
-import org.spongepowered.configurate.CommentedConfigurationNode;
 import org.spongepowered.configurate.ConfigurateException;
-import org.spongepowered.configurate.reference.ValueReference;
 import org.spongepowered.math.vector.Vector3i;
 
-import io.leangen.geantyref.TypeToken;
-
-import net.kyori.adventure.text.Component;
-
 import sawfowl.commandpack.utils.StorageType;
-import sawfowl.localeapi.api.serializetools.SerializeOptions;
 import sawfowl.regionguard.RegionGuard;
-import sawfowl.regionguard.api.RegionSerializerCollection;
 import sawfowl.regionguard.api.RegionTypes;
 import sawfowl.regionguard.api.SelectorTypes;
-import sawfowl.regionguard.api.data.AdditionalData;
-import sawfowl.regionguard.api.data.AdditionalDataMap;
 import sawfowl.regionguard.api.data.ClaimedByPlayer;
 import sawfowl.regionguard.api.data.Cuboid;
-import sawfowl.regionguard.api.data.FlagValue;
-import sawfowl.regionguard.api.data.MemberData;
 import sawfowl.regionguard.api.data.PlayerData;
 import sawfowl.regionguard.api.data.PlayerLimits;
 import sawfowl.regionguard.api.data.Region;
 import sawfowl.regionguard.configure.MySQL;
-import sawfowl.regionguard.data.PlayerDataImpl;
-import sawfowl.regionguard.data.RegionImpl;
 
 public class MySqlStorage extends AbstractSqlStorage {
 
 	private final MySQL mySQL;
 	private Statement statement;
 	private ScheduledTask sync;
-	private final TypeToken<Map<String, Component>> mapComponentsToken = new TypeToken<Map<String, Component>>(){};
-	private final TypeToken<Map<String, Set<FlagValue>>> flagsToken = new TypeToken<Map<String, Set<FlagValue>>>(){};
-	private final TypeToken<List<MemberData>> membersToken = new TypeToken<List<MemberData>>(){};
-	private final TypeToken<AdditionalDataMap<? extends AdditionalData>> dataMapToken = new TypeToken<AdditionalDataMap<?>>(){};
 	private String lastGlobalSync;
+	private String lastRegionSync;
+	private String lastPlayerSync;
 	public MySqlStorage(RegionGuard plugin) {
 		super(plugin);
 		this.mySQL = plugin.getMySQL();
@@ -74,7 +55,7 @@ public class MySqlStorage extends AbstractSqlStorage {
 			createWorldsTables();
 			createDataForWorlds();
 		}
-		sync = sync();
+		sync = syncTask();
 		Sponge.eventManager().registerListeners(plugin.getPluginContainer(), this);
 	}
 
@@ -83,52 +64,29 @@ public class MySqlStorage extends AbstractSqlStorage {
 		return statement == null || statement.isClosed() ? statement = plugin.getMySQL().getOrOpenConnection().createStatement() : statement;
 	}
 
-	public void createTableForPlayers() {
-		executeSQL("CREATE TABLE IF NOT EXISTS " + prefix + "player_data(uuid VARCHAR(128) UNIQUE, claimed_blocks BIGINT, claimed_regions BIGINT, limit_blocks BIGINT, limit_regions BIGINT, limit_subdivisions BIGINT, limit_members BIGINT, written DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP, PRIMARY KEY(uuid));");
-	}
-
-	public void createWorldsTables() {
-		executeSQL("CREATE TABLE IF NOT EXISTS " + prefix + "worlds(uuid VARCHAR(128) UNIQUE, world VARCHAR(128) UNIQUE, name LONGTEXT, region_type TEXT, creation_time BIGINT, join_message LONGTEXT, exit_message LONGTEXT, flags LONGTEXT, additional_data LONGTEXT, written DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP, PRIMARY KEY(world));");
-		for(ResourceKey worldKey : Sponge.server().worldManager().worldKeys()) {
-			executeSQL("CREATE TABLE IF NOT EXISTS " + prefix + "world_" + worldKey.asString().replace(':', '_')  + "(uuid VARCHAR(128) UNIQUE, name LONGTEXT, region_type TEXT, creation_time BIGINT, join_message LONGTEXT, exit_message LONGTEXT, flags LONGTEXT, members LONGTEXT, min_x BIGINT, min_y BIGINT, min_z BIGINT, max_x BIGINT, max_y BIGINT, max_z BIGINT, selector_type TEXT, additional_data LONGTEXT, parrent VARCHAR(128), written DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP, PRIMARY KEY(uuid));");
-		}
-	}
-
-	@Override
-	public void createDataForWorlds() {
-		Sponge.server().worldManager().worlds().forEach(world -> {
-			Region global = getWorldRegion(world);
-			if(global == null) {
-				global = Region.createGlobal(world, plugin.getDefaultFlagsConfig().getGlobalFlags());
-				saveRegion(global);
-				plugin.getAPI().updateGlobalRegionData(world, global);
-			} else {
-				plugin.getAPI().updateGlobalRegionData(world, global);
-			}
-		});
-	}
-
 	@Override
 	public Region getWorldRegion(ServerWorld world) {
 		try {
 			ResultSet results = resultSet("SELECT * FROM " + prefix + "worlds;");
 			while(!results.isClosed() && results.next()) {
 				if(world.key().asString().equals(results.getString("world"))) {
-					lastGlobalSync = results.getString("written");
+					if(lastGlobalSync == null) lastGlobalSync = results.getString("written");
 					return getGlobalRegionfromResultSet(results, world);
 				}
 			}
 		} catch (SQLException | ConfigurateException e) {
 			plugin.getLogger().error("Get global region data. World " + world.key().asString() + "\n" + e.getLocalizedMessage());
 		}
-		return globalsCreated ? (RegionImpl) Region.createGlobal(world, plugin.getDefaultFlagsConfig().getGlobalFlags()) : null;
+		Region region = Region.createGlobal(world, plugin.getDefaultFlagsConfig().getGlobalFlags());
+		saveRegion(region);
+		return region;
 	}
 
 	@Override
 	public void saveRegion(Region region) {
 		String sql = null;
 		if(region.isGlobal()) {
-			sql = "REPLACE INTO " + prefix + "worlds(uuid, world, name, region_type, creation_time, join_message, exit_message, flags, additional_data) VALUES('"
+			sql = "REPLACE INTO " + prefix + "worlds(uuid, world, name, region_type, creation_time, join_message, exit_message, flags, members, additional_data) VALUES('"
 			+ region.getUniqueId().toString() + "', '"
 			+ region.getWorldKey().asString() + "', '"
 			+ getSerializedData(region.getNames(), mapComponentsToken) + "', '"
@@ -137,6 +95,7 @@ public class MySqlStorage extends AbstractSqlStorage {
 			+ getSerializedData(region.getJoinMessages(), mapComponentsToken) + "', '"
 			+ getSerializedData(region.getExitMessages(), mapComponentsToken) + "', '"
 			+ getSerializedData(region.getFlags(), flagsToken) + "', '"
+			+ getSerializedData(region.getMembers(), membersToken) + "', '"
 			+ getSerializedData(region.getAllAdditionalData(), dataMapToken) + "', '"
 			+ region.getWorldKey().asString()
 			+ "');";
@@ -166,6 +125,7 @@ public class MySqlStorage extends AbstractSqlStorage {
 
 	@Override
 	public void deleteRegion(Region region) {
+		if(region.isGlobal()) return;
 		executeSQL("DELETE FROM " + prefix + "world_" + region.getWorldKey().asString().replace(':', '_') + " WHERE " + prefix + "world_" + region.getWorldKey().asString().replace(':', '_') + ".uuid = '" + region.getUniqueId().toString() + "';");
 		if(!region.getChilds().isEmpty()) region.getChilds().forEach(this::deleteRegion);
 	}
@@ -175,32 +135,29 @@ public class MySqlStorage extends AbstractSqlStorage {
 		Sponge.server().worldManager().worlds().forEach(world -> {
 			Map<UUID, Set<Region>> childs = new HashMap<>();
 			try {
-				ResultSet results = resultSet("SELECT * FROM " + prefix + "world_" + world.key().asString().replace(':', '_') + ";");
+				ResultSet results = resultSet("SELECT * FROM " + prefix + "world_" + world.key().asString().replace(':', '_') + " ORDER BY written;");
 				while(!results.isClosed() && results.next()) {
-					UUID uuid = UUID.fromString(results.getString("uuid"));
-					UUID parrent = UUID.fromString(results.getString("parrent"));
+					if(lastRegionSync == null) lastRegionSync =  results.getString("written");
 					Region region = getRegionfromResultSet(results, world);
+					UUID uuid = region.getUniqueId();
+					String parrent = results.getString("parrent");
 					if(parrent != null) {
-						if(plugin.getAPI().getRegions().stream().filter(rg -> rg.getUniqueId().equals(parrent)).findFirst().isPresent()) {
-							plugin.getAPI().getRegions().stream().filter(rg -> rg.getUniqueId().equals(parrent)).findFirst().get().addChild(region);
+						if(plugin.getAPI().getRegions().stream().filter(rg -> rg.getUniqueId().toString().equals(parrent)).findFirst().isPresent()) {
+							plugin.getAPI().getRegions().stream().filter(rg -> rg.getUniqueId().toString().equals(parrent)).findFirst().get().addChild(region);
 						} else {
-							if(!childs.containsKey(parrent)) childs.put(parrent, new HashSet<Region>());
-							childs.get(parrent).add(region);
+							if(!childs.containsKey(UUID.fromString(parrent))) childs.put(UUID.fromString(parrent), new HashSet<Region>());
+							childs.get(UUID.fromString(parrent)).add(region);
 						}
 					} else {
 						if(childs.containsKey(uuid)) childs.get(uuid).forEach(child -> region.addChild(region));
 						plugin.getAPI().registerRegion(region);
 					}
 				}
+				plugin.getAPI().updateGlobalRegionData(world, getWorldRegion(world));
 			} catch (SQLException | ConfigurateException e) {
 				plugin.getLogger().error("Load region data\n" + e.getLocalizedMessage());
 			}
 		});
-	}
-
-	@Override
-	public void savePlayerData(ServerPlayer player, PlayerData playerData) {
-		savePlayerData(player.uniqueId(), playerData);
 	}
 
 	@Override
@@ -224,14 +181,15 @@ public class MySqlStorage extends AbstractSqlStorage {
 		} catch (SQLException e) {
 			plugin.getLogger().error("Get player data. Player: " + player.name() + "\n" + e.getLocalizedMessage());
 		}
-		return (PlayerDataImpl) PlayerData.zero();
+		return PlayerData.zero();
 	}
 
 	@Override
 	public void loadDataOfPlayers() {
 		try {
-			ResultSet results = resultSet("SELECT * FROM " + prefix + "player_data");
+			ResultSet results = resultSet("SELECT * FROM " + prefix + "player_data ORDER BY written");
 			while(!results.isClosed() && results.next()) {
+				if(this.lastPlayerSync == null) this.lastPlayerSync = results.getString("written");
 				plugin.getAPI().setPlayerData(UUID.fromString(results.getString("uuid")), getPlayerDataFromResultSet(results));
 			}
 		} catch (SQLException e) {
@@ -242,14 +200,26 @@ public class MySqlStorage extends AbstractSqlStorage {
 	@Listener
 	public void onRefresh(RefreshGameEvent event) {
 		if(!sync.isCancelled()) sync.cancel();
-		sync = sync();
+		sync = syncTask();
 	}
 
-	private ScheduledTask sync() {
+	private void createTableForPlayers() {
+		executeSQL("CREATE TABLE IF NOT EXISTS " + prefix + "player_data(uuid VARCHAR(128) UNIQUE, claimed_blocks BIGINT, claimed_regions BIGINT, limit_blocks BIGINT, limit_regions BIGINT, limit_subdivisions BIGINT, limit_members BIGINT, written DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP, PRIMARY KEY(uuid));");
+	}
+
+	private void createWorldsTables() {
+		executeSQL("CREATE TABLE IF NOT EXISTS " + prefix + "worlds(uuid VARCHAR(128) UNIQUE, world VARCHAR(128) UNIQUE, name LONGTEXT, region_type TEXT, creation_time BIGINT, join_message LONGTEXT, exit_message LONGTEXT, flags LONGTEXT, members LONGTEXT, additional_data LONGTEXT, written DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP, PRIMARY KEY(world));");
+		for(ResourceKey worldKey : Sponge.server().worldManager().worldKeys()) {
+			executeSQL("CREATE TABLE IF NOT EXISTS " + prefix + "world_" + worldKey.asString().replace(':', '_')  + "(uuid VARCHAR(128) UNIQUE, name LONGTEXT, region_type TEXT, creation_time BIGINT, join_message LONGTEXT, exit_message LONGTEXT, flags LONGTEXT, members LONGTEXT, min_x BIGINT, min_y BIGINT, min_z BIGINT, max_x BIGINT, max_y BIGINT, max_z BIGINT, selector_type TEXT, additional_data LONGTEXT, parrent VARCHAR(128), written DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP, PRIMARY KEY(uuid));");
+		}
+	}
+
+	private ScheduledTask syncTask() {
 		return Sponge.asyncScheduler().submit(Task.builder().plugin(plugin.getPluginContainer()).interval(plugin.getConfig().getMySQLConfig().getSyncInterval(), TimeUnit.SECONDS).execute(() -> {
 			for(ServerWorld world : Sponge.server().worldManager().worlds()) {
 				try {
-					syncLoad(world);
+					syncRegions(world);
+					syncPlayers();
 				} catch (SQLException | ConfigurateException e) {
 					plugin.getLogger().error(e.getLocalizedMessage());
 				}
@@ -257,38 +227,63 @@ public class MySqlStorage extends AbstractSqlStorage {
 		}).build());
 	}
 
-	//private void syncLoad() {
-		
-	//}
-
-	private void syncLoad(ServerWorld world) throws SQLException, ConfigurateException {
-		ResultSet globalSet = resultSet(this.lastGlobalSync == null ? "SELECT * FROM " + prefix + "worlds ORDER BY written DESC LIMIT 1;" : "SELECT * FROM " + prefix + "worlds WHERE written > '" + this.lastGlobalSync + "' ORDER BY written;");
-		boolean updateTimeGlobal = false;
-		while(!globalSet.isClosed() && globalSet.next()) {
-			String key = globalSet.getString("world");
-			if(key.equals(world.key().asString())) {
-				if(lastGlobalSync == null || !updateTimeGlobal) {
-					lastGlobalSync = globalSet.getString("written");
-					updateTimeGlobal = true;
+	private void syncPlayers() throws SQLException {
+		if(this.lastPlayerSync != null) {
+			ResultSet playersSet = resultSet("SELECT * FROM " + prefix + "player_data WHERE written > '" + this.lastPlayerSync + "' ORDER BY written;");
+			boolean updateTimePlayers = false;
+			while(!playersSet.isClosed() && playersSet.next()) {
+				if(!updateTimePlayers) {
+					lastPlayerSync = playersSet.getString("written");
+					updateTimePlayers = true;
 				}
-				plugin.getAPI().updateGlobalRegionData(world, getGlobalRegionfromResultSet(globalSet, world));
+				UUID uuid = UUID.fromString(playersSet.getString("uuid"));
+				Optional<PlayerData> optData = plugin.getAPI().getPlayerData(uuid);
+				if(optData.isPresent()) {
+					optData.get().setLimits(getPlayerLimits(playersSet)).setClaimed(getClaimedByPlayer(playersSet));
+				} else plugin.getAPI().setPlayerData(uuid, getPlayerDataFromResultSet(playersSet));
 			}
-		}
+		} else loadDataOfPlayers();
 	}
 
-	private <T> ValueReference<T, CommentedConfigurationNode> createTempConfigReader(String string, TypeToken<T> token) throws ConfigurateException {
-		return SerializeOptions.createHoconConfigurationLoader(2).defaultOptions(options -> options.serializers(serializers -> serializers.registerAll(RegionSerializerCollection.COLLETCTION))).source(() -> new BufferedReader(new StringReader(string))).build().loadToReference().referenceTo(token);
-	}
-
-	private <T> String getSerializedData(T object, TypeToken<T> token) {
-		StringWriter sink = new StringWriter();
-		try {
-			SerializeOptions.createHoconConfigurationLoader(2).defaultOptions(options -> options.serializers(serializers -> serializers.registerAll(RegionSerializerCollection.COLLETCTION))).sink(() -> new BufferedWriter(sink)).build().loadToReference().referenceTo(token).setAndSave(object);
-			return sink.toString();
-		} catch (ConfigurateException e) {
-			e.printStackTrace();
-		}
-		return null;
+	private void syncRegions(ServerWorld world) throws SQLException, ConfigurateException {
+		if(this.lastGlobalSync != null) {
+			ResultSet globalSet = resultSet("SELECT * FROM " + prefix + "worlds WHERE written > '" + this.lastGlobalSync + "' ORDER BY written;");
+			boolean updateTimeGlobal = false;
+			while(!globalSet.isClosed() && globalSet.next()) {
+				String key = globalSet.getString("world");
+				if(key.equals(world.key().asString())) {
+					if(!updateTimeGlobal) {
+						lastGlobalSync = globalSet.getString("written");
+						updateTimeGlobal = true;
+					}
+					plugin.getAPI().updateGlobalRegionData(world, getGlobalRegionfromResultSet(globalSet, world));
+				}
+			}
+		} else plugin.getAPI().updateGlobalRegionData(world, getWorldRegion(world));
+		if(this.lastRegionSync != null) {
+			ResultSet regionsSet = resultSet("SELECT * FROM " + prefix + "world_" + world.key().asString().replace(':', '_') + " WHERE written > '" + this.lastRegionSync + "' ORDER BY written;");
+			boolean updateTimeRegion = false;
+			while(!regionsSet.isClosed() && regionsSet.next()) {
+				if(!updateTimeRegion) {
+					lastRegionSync = regionsSet.getString("written");
+					updateTimeRegion = true;
+				}
+				Region region = getRegionfromResultSet(regionsSet, world);
+				UUID uuid = region.getUniqueId();
+				String parrent = regionsSet.getString("parrent");
+				if(parrent == null) {
+					Optional<Region> registered = plugin.getAPI().getRegions().stream().filter(rg -> rg.getUniqueId().equals(uuid)).findFirst();
+					if(registered.isPresent()) plugin.getAPI().unregisterRegion(registered.get());
+					plugin.getAPI().registerRegion(region);
+				} else {
+					Optional<Region> findParrent = plugin.getAPI().getRegions().stream().map(rg -> rg.getAllChilds()).flatMap(Collection::stream).filter(rg -> rg.getParrent().isPresent() && rg.getParrent().get().getUniqueId().toString().equals(parrent)).findFirst();
+					if(findParrent.isPresent()) {
+						findParrent.get().getChilds().removeIf(child -> child.getUniqueId().equals(uuid));
+						findParrent.get().addChild(region);
+					}
+				}
+			}
+		} else loadRegions();
 	}
 
 	private ClaimedByPlayer getClaimedByPlayer(ResultSet results) throws SQLException {
@@ -332,6 +327,7 @@ public class MySqlStorage extends AbstractSqlStorage {
 		String joinMessage = results.getString("join_message");
 		String exitMessage = results.getString("exit_message");
 		String flags = results.getString("flags");
+		String members = results.getString("members");
 		String additionalData = results.getString("additional_data");
 		return Region.builder()
 				.setUniqueId(UUID.fromString(results.getString("uuid")))
@@ -341,6 +337,7 @@ public class MySqlStorage extends AbstractSqlStorage {
 				.addJoinMessages(joinMessage != null && !joinMessage.isEmpty() ? createTempConfigReader(joinMessage, mapComponentsToken).get() : null)
 				.addExitMessages(exitMessage != null && !exitMessage.isEmpty() ? createTempConfigReader(exitMessage, mapComponentsToken).get() : null)
 				.setFlags(flags != null && !flags.isEmpty() ? createTempConfigReader(flags, flagsToken).get() : null)
+				.addMembers(members != null && !members.isEmpty() ? createTempConfigReader(members, membersToken).get() : null)
 				.addAdditionalData(additionalData != null && !additionalData.isEmpty() ? createTempConfigReader(additionalData, dataMapToken).get() : null)
 				.setServerOwner()
 				.setType(RegionTypes.GLOBAL)
